castling make sure that a king does not pass through or finish on a square that is attacked by an enemy.

+++ src/main/java/com/stateofflux/chess/model/pieces/StraightLineMoves.java
@@ -24,7 +24,7 @@ public abstract class StraightLineMoves extends PieceMoves {
         // validation here - throw IllegalArgumentException with details when invalid
         int nextPosition;
         long nextPositionBit;
-        long boardMax;
+        int boardMax;
 
         // calculate the max moves
         for (Direction d : this.directions) {
